pyproject.toml
[tool.poetry]
name = "lms-drf"
version = "0.1.0"
description = ""
authors = ["Tsurkan <mikh.tsurkan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
djangorestframework = "^3.15.1"
psycopg2 = "^2.9.9"
pillow = "^10.3.0"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
docutils = "^0.21.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
